refactor(@appcraft/mui): separating the management of context and custom hook
