+++ libs/mui/src/components/TypeList/TypeList.tsx
@@ -2,11 +2,11 @@ import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 import List from '@mui/material/List';
 import PlaylistAddIcon from '@mui/icons-material/PlaylistAdd';
 
+import * as Hooks from '../../hooks';
 import * as Styles from '../../styles';
 import { Breadcrumbs } from '../common';
-import { OptionValues, useFixedT } from '../../contexts';
 import { TypeItem } from '../TypeItem';
-import { usePropertyRouter, useTypeItems } from '../../hooks';
+import type { OptionValues } from '../../contexts';
 import type { TypeListProps } from './TypeList.types';
 
 export default function TypeList<V extends OptionValues>({
@@ -16,13 +16,17 @@ export default function TypeList<V extends OptionValues>({
   onChange,
   onCollectionPathChange,
 }: TypeListProps<V>) {
-  const ct = useFixedT();
-  const { items, onItemAdd } = useTypeItems<V>(superior, values, onChange);
-  const isModifiable = onItemAdd instanceof Function;
+  const [breadcrumbs, { back: handleBack, to: handleTo }] =
+    Hooks.usePropertyRouter(onCollectionPathChange);
 
-  const [breadcrumbs, { back: handleBack, to: handleTo }] = usePropertyRouter(
-    onCollectionPathChange
+  const { items, onItemAdd } = Hooks.useTypeItems<V>(
+    superior,
+    values,
+    onChange
   );
+
+  const ct = Hooks.useFixedT();
+  const isModifiable = onItemAdd instanceof Function;
 
   return (
     <List
