+++ libs/mui/src/contexts/Editor/index.ts
@@ -1,0 +1,9 @@
+export { EditorProvider, useEditorContext } from './Editor';
+
+export type {
+  ChangeHandler,
+  FixedT,
+  OptionValues,
+  EditorContextValue,
+  EditorProviderProps,
+} from './Editor.types';
