--- libs/mui/src/contexts/EditorContext/index.ts
@@ -1,17 +1,0 @@
-export { default as EditorProvider } from './EditorProvider';
-
-export {
-  useCollection,
-  useFixedT,
-  useMixedTypeMapping,
-  usePropValue,
-} from './EditorContext.hooks';
-
-export type {
-  ChangeHandler,
-  Collection,
-  EditorProviderProps,
-  FixedT,
-  MixedTypeMappingResult,
-  OptionValues,
-} from './EditorContext.types';
