+++ libs/mui/src/contexts/Renderer/index.ts
@@ -1,0 +1,2 @@
+export { RendererProvider, useRendererContext } from './Renderer';
+export type { RendererProviderProps } from './Renderer.types';
