--- libs/mui/src/contexts/RendererContext/RendererContext.hooks.tsx
@@ -1,24 +1,0 @@
-import { createContext, lazy, useContext, useMemo } from 'react';
-import type * as Appcraft from '@appcraft/types';
-
-import type * as Types from './RendererContext.types';
-
-export const RendererContext = createContext<Types.RendererContextValue>({});
-
-export const useLazyWidget = (widget: Appcraft.WidgetOptions) => {
-  const { type } = widget as Appcraft.NodeWidget;
-
-  const { lazy: externalLazy } = useContext(
-    RendererContext
-  ) as Required<Types.RendererContextValue>;
-
-  return useMemo(
-    () =>
-      type
-        ? externalLazy(type)
-        : lazy(async () => ({
-            default: ({ children }) => children,
-          })),
-    [externalLazy, type]
-  );
-};
