--- libs/mui/src/contexts/RendererContext/index.ts
@@ -1,3 +1,0 @@
-export { default as RendererProvider } from './RendererProvider';
-export { useLazyWidget } from './RendererContext.hooks';
-export type { RendererProviderProps } from './RendererContext.types';
