+++ libs/mui/src/contexts/index.ts
@@ -1,2 +1,2 @@
-export * from './EditorContext';
-export * from './RendererContext';
+export * from './Editor';
+export * from './Renderer';
