+++ libs/mui/src/hooks/index.ts
@@ -1,6 +1,11 @@
+export * from './useCollection';
 export * from './useDisplayPropName';
+export * from './useFixedT';
 export * from './useLazyTypeList';
+export * from './useLazyWidget';
 export * from './useLazyWidgetNodes';
+export * from './useMixedTypeMapping';
+export * from './usePropValue';
 export * from './usePropertiesSorting';
 export * from './usePropertyRouter';
 export * from './useStructure';
