+++ libs/mui/src/hooks/useCollection/index.ts
@@ -1,0 +1,1 @@
+export { default as useCollection } from './useCollection';
