+++ libs/mui/src/hooks/useCollection/useCollection.types.ts
@@ -1,0 +1,1 @@
+export type Collection = Array<unknown> | Record<string, unknown>;
