+++ libs/mui/src/hooks/useFixedT/index.ts
@@ -1,0 +1,2 @@
+export { default as useFixedT } from './useFixedT';
+export type { FixedT } from './useFixedT.types';
