+++ libs/mui/src/hooks/useFixedT/useFixedT.ts
@@ -1,0 +1,35 @@
+import { useMemo } from 'react';
+
+import { useEditorContext } from '../../contexts';
+import type * as Types from './useFixedT.types';
+
+const replaces: Types.Replaces = [
+  { pattern: /^[^-]+-/, replacement: '' },
+  {
+    pattern: /-(.)/g,
+    replacement: (_match, letter) => ` ${letter.toUpperCase()}`,
+  },
+  {
+    pattern: /^./,
+    replacement: (match) => match.toUpperCase(),
+  },
+];
+
+const useFixedT: Types.FixedTHook = (defaultFixedT) => {
+  const { fixedT } = useEditorContext();
+
+  return useMemo<Types.FixedT>(
+    () =>
+      fixedT ||
+      defaultFixedT ||
+      ((key) =>
+        replaces.reduce<string>(
+          (result, { pattern, replacement }) =>
+            result.replace(pattern, replacement as string),
+          key
+        )),
+    [defaultFixedT, fixedT]
+  );
+};
+
+export default useFixedT;
