+++ libs/mui/src/hooks/useLazyWidget/index.ts
@@ -1,0 +1,1 @@
+export { default as useLazyWidget } from './useLazyWidget';
