+++ libs/mui/src/hooks/useMixedTypeMapping/index.ts
@@ -1,0 +1,1 @@
+export { default as useMixedTypeMapping } from './useMixedTypeMapping';
