+++ libs/mui/src/hooks/useMixedTypeMapping/useMixedTypeMapping.types.ts
@@ -1,0 +1,4 @@
+export type MixedTypeMappingResult = [
+  string | null,
+  (mixedText?: string) => void
+];
