+++ libs/mui/src/hooks/usePropValue/index.ts
@@ -1,0 +1,1 @@
+export { default as usePropValue } from './usePropValue';
